+++ packages/admin-stories/src/admin/stack/Stack.tsx
@@ -1,17 +1,12 @@
-import { Stack, StackBreadcrumbsContainer } from "@comet/admin";
+import { Stack } from "@comet/admin";
 import { Typography } from "@material-ui/core";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
 import StoryRouter from "storybook-react-router";
-import styled from "styled-components";
-
-const BreadcrumbsContainer = styled(StackBreadcrumbsContainer)`
-    border: 1px solid red;
-`;
 
 function Story() {
     return (
-        <Stack topLevelTitle="Stack" components={{ breadcrumbsContainer: BreadcrumbsContainer }}>
+        <Stack topLevelTitle="Stack">
             <Typography>Foo</Typography>
         </Stack>
     );