+++ packages/admin/src/form/FormPaper.tsx
@@ -1,6 +1,20 @@
-import { Paper } from "@material-ui/core";
-import styled from "styled-components";
+import { Paper, PaperProps } from "@material-ui/core";
+import { Theme } from "@material-ui/core/styles";
+import { makeStyles } from "@material-ui/styles";
+import * as React from "react";
 
-export const FormPaper = styled(Paper)`
-    padding: 16px;
-`;
+export type CometAdminFormPaperKeys = "root";
+
+const useStyles = makeStyles<Theme, {}, CometAdminFormPaperKeys>(
+    () => ({
+        root: {
+            padding: 16,
+        },
+    }),
+    { name: "CometAdminFormPaper" },
+);
+
+export function FormPaper(props: PaperProps): React.ReactElement {
+    const classes = useStyles();
+    return <Paper classes={{ root: classes.root }} {...props} />;
+}